+++ Changelog.md
@@ -3,6 +3,7 @@ Changes:
 
 ## Version 0.4.3
 - Updated `simd` dependency to version `0.2`.
+- New feature: Added `ApproximatedOps` trait.
 
 ## Version 0.4.2
 Bugfix release.